entry: handle keyboard events in unix part
