+++ unix/uipriv_unix.h
@@ -1,4 +1,5 @@
 // 22 april 2015
+#pragma once
 #define GLIB_VERSION_MIN_REQUIRED GLIB_VERSION_2_40
 #define GLIB_VERSION_MAX_ALLOWED GLIB_VERSION_2_40
 #define GDK_VERSION_MIN_REQUIRED GDK_VERSION_3_10
